feat: 구간이 1개인 경우에 역을 제거하려면 오류가 발생하는 테스트 및 기능 추가

+++ src/main/java/nextstep/subway/applicaion/SectionService.java
@@ -4,10 +4,7 @@ import lombok.RequiredArgsConstructor;
 import nextstep.subway.applicaion.dto.section.SectionRequest;
 import nextstep.subway.domain.Section;
 import nextstep.subway.domain.SubwayLine;
-import nextstep.subway.exception.AlreadyRegisterException;
-import nextstep.subway.exception.CannotDeleteException;
-import nextstep.subway.exception.ErrorCode;
-import nextstep.subway.exception.SameUpStationException;
+import nextstep.subway.exception.*;
 import nextstep.subway.repository.SectionRepository;
 import nextstep.subway.repository.SubwayLineRepository;
 import org.springframework.stereotype.Service;
@@ -42,12 +39,20 @@ public class SectionService {
 	public void delete(Long subwayLineId, Long stationId) {
 		SubwayLine subwayLine = subwayLineRepository.findById(subwayLineId).orElseThrow(NoSuchElementException::new);
 
+		if (isSectionSizeEqualsOne(subwayLine)) {
+			throw new CannotDeleteUniqueSectionException(ErrorCode.CANNOT_DELETE_UNIQUE_SECTION.getMessage());
+		}
+
 		if (isNotSameDownStationId(stationId, subwayLine)) {
 			throw new CannotDeleteException(ErrorCode.CANNOT_DELETE_WITH_NOT_SAME_DOWN_STATION.getMessage());
 		}
 
 		Section deleteSection = subwayLine.deleteSection(stationId);
 		sectionRepository.delete(deleteSection);
+	}
+
+	private boolean isSectionSizeEqualsOne(SubwayLine subwayLine) {
+		return subwayLine.getSectionList().size() == 1;
 	}
 
 	private boolean isNotSameDownStationId(Long stationId, SubwayLine subwayLine) {
